Move add_image_to_doc function location
