+++ aust_covid/model.py
@@ -12,7 +12,7 @@ from summer2.parameters import Parameter, Function, Time
 
 from emutools.tex import StandardTexDoc, get_tex_formatted_date
 from emutools.utils import capture_kwargs
-from emutools.utils import triangle_wave_func, add_image_to_doc
+from emutools.tex import triangle_wave_func, add_image_to_doc
 from inputs.constants import REFERENCE_DATE, ANALYSIS_START_DATE, ANALYSIS_END_DATE, WA_REOPEN_DATE, MATRIX_LOCATIONS
 from inputs.constants import N_LATENT_COMPARTMENTS, AGE_STRATA, STRAIN_STRATA, INFECTION_PROCESSES, DATA_PATH
 from aust_covid.inputs import load_pop_data, load_uk_pop_data, get_base_vacc_data
