+++ emutools/utils.py
@@ -1,14 +1,10 @@
-from pathlib import Path
-from typing import Union
 from jax import numpy as jnp
 from jax import scipy as jsp
 import numpy as np
-from matplotlib.figure import Figure as MplFig
-from plotly.graph_objects import Figure as PlotlyFig
 
-from emutools.tex import StandardTexDoc
 from summer2.parameters import Function, Data, DerivedOutput
 
+from emutools.tex import StandardTexDoc
 from inputs.constants import SUPPLEMENT_PATH
 
 
@@ -94,44 +90,6 @@ def build_gamma_dens_interval_func(
     lags = Data(model_times - model_times[0])
     cdf_values = Function(gamma_cdf, [shape, mean, lags])
     return Function(jnp.gradient, [cdf_values])
-
-
-def add_image_to_doc(
-    fig: Union[MplFig, PlotlyFig], 
-    filename: str, 
-    filetype: str,
-    title: str, 
-    tex_doc: StandardTexDoc, 
-    section: str,
-    subsection: str='',
-    caption: str='',
-    fig_width: float=0.85,
-):
-    """
-    Save a figure image to a local directory and include in TeX doc.
-
-    Args:
-        fig: The figure object
-        filename: A string for the filenam to save the figure as
-        caption: Figure caption for the document
-        tex_doc: The working document
-        section: Section of the document to include the figure in
-
-    Raises:
-        TypeError: If the figure is not one of the two supported formats
-    """
-    fig_folder = 'figures'
-    fig_path = SUPPLEMENT_PATH / fig_folder
-    Path(fig_path).mkdir(exist_ok=True)
-    full_filename = f'{filename}.{filetype}'
-    if isinstance(fig, MplFig):
-        fig.savefig(fig_path / full_filename)
-    elif isinstance(fig, PlotlyFig):
-        fig.update_layout(margin={'t': 25})
-        fig.write_image(fig_path / full_filename)
-    else:
-        raise TypeError('Figure type not supported')
-    tex_doc.include_figure(title, filename, filetype, fig_folder, section, subsection=subsection, caption=caption, fig_width=fig_width)
 
 
 def capture_kwargs(*args, **kwargs):
